Feature/11 postgresql migrations

+++ src/Mapingway.Domain/Auth/RefreshToken.cs
@@ -2,11 +2,11 @@
 
 public class RefreshToken
 {
-    public int Id { get; init; }
+    public long Id { get; init; }
     public string Value { get; init; } = null!;
-    public int? UserId { get; init; }
+    public long? UserId { get; init; }
     public User? User { get; set; }
-    public int? TokenFamilyId { get; init; }
+    public long? TokenFamilyId { get; init; }
     public RefreshTokenFamily? TokenFamily { get; init; }
     public bool IsUsed { get; set; }
     public DateTime ExpiresAt { get; init; }
