+++ src/Mapingway.Domain/Auth/UserRole.cs
@@ -2,7 +2,7 @@
 
 public class UserRole
 {
-    public int UserId { get; init; }
+    public long UserId { get; init; }
     public User User { get; init; } = null!;
     public int RoleId { get; init; }
     public Role Role { get; init; } = null!;
